+++ R/prepInputs.R
@@ -300,8 +300,12 @@ prepInputs <- function(targetFile = NULL, url = NULL, archive = NULL, alsoExtrac
         if (returnAsList)
           as.list(tmpEnv, all.names = TRUE)
       } else {
-        Cache(do.call, out$fun, append(list(asPath(out$targetFilePath)), args),
-              useCache = useCache)
+        mess <- capture.output(type = "message", out <- Cache(do.call, out$fun, append(list(asPath(out$targetFilePath)), args),
+              useCache = useCache))
+        mess <- grep("No cacheRepo supplied", mess, invert = TRUE, value = TRUE)
+        if (length(mess) > 0)
+          message(mess)
+        return(out)
       }
     }
   } else {
@@ -741,7 +745,7 @@ appendChecksumsTable <- function(checkSumFilePath, filesToChecksum,
           "  what to look for.")
   capture.output(
     type = "message",
-    currentFiles <- Checksums(path = destinationPath, write = TRUE,
+    currentFiles <- Checksums(path = destinationPath, write = !append,
                               files = file.path(destinationPath, filesToChecksum))
   )
   if (append) { # a checksums file already existed, need to keep some of it