+++ inst/examples/example_Cache.R
@@ -9,7 +9,7 @@ ranNumsC <- rnorm(10, 16) %>% Cache(cacheRepo = tmpDir) # recovers cached copy
 ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy
 
 # experimental devMode
-opt <- options("reproducible.devMode" = TRUE)
+opt <- options("reproducible.useCache" = "devMode")
 clearCache(tmpDir, ask = FALSE)
 centralTendency <- function(x)
   mean(x)